Remove AGP 8 compatibility shims for non-transitive R and non-final res IDs
Convert all switch/case on R.id and R.string to if/else chains across
ActivityMain, FragmentPlayerSmall, PlayerSelectorAdapter, and
StationActions. Fix transitive R class references to use library-owned
R classes directly: androidx.appcompat.R.attr.colorAccent in Utils,
com.google.android.material.R.attr.boxBackgroundColor in
ItemAdapterStation and ItemAdapterIconOnlyStation, and
com.github.zawadz88.materialpopupmenu.R.style in StationPopupMenu.

Remove android.nonTransitiveRClass=false and android.nonFinalResIds=false
from gradle.properties, adopting AGP 8 defaults.

+++ app/src/main/java/net/programmierecke/radiodroid2/FragmentPlayerSmall.java
@@ -286,23 +286,15 @@ public class FragmentPlayerSmall extends Fragment {
         final PopupMenu dropDownMenu = new PopupMenu(getContext(), buttonMore);
         dropDownMenu.getMenuInflater().inflate(R.menu.menu_player, dropDownMenu.getMenu());
         dropDownMenu.setOnMenuItemClickListener(menuItem -> {
-            switch (menuItem.getItemId()) {
-                case R.id.action_homepage: {
-                    StationActions.showWebLinks(requireActivity(), currentStation);
-                    break;
-                }
-                case R.id.action_share: {
-                    StationActions.share(requireContext(), currentStation);
-                    break;
-                }
-                case R.id.action_set_alarm: {
-                    StationActions.setAsAlarm(requireActivity(), currentStation);
-                    break;
-                }
-                case R.id.action_delete_stream_history: {
-                    trackHistoryRepository.deleteHistory();
-                    break;
-                }
+            int itemId = menuItem.getItemId();
+            if (itemId == R.id.action_homepage) {
+                StationActions.showWebLinks(requireActivity(), currentStation);
+            } else if (itemId == R.id.action_share) {
+                StationActions.share(requireContext(), currentStation);
+            } else if (itemId == R.id.action_set_alarm) {
+                StationActions.setAsAlarm(requireActivity(), currentStation);
+            } else if (itemId == R.id.action_delete_stream_history) {
+                trackHistoryRepository.deleteHistory();
             }
 
             return true;
